src/requirements.txt
requests
beautifulsoup4
readability-lxml
pydantic-ai